requirements.txt
python-memcached==1.62
